# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OF.ru",
        "version": "1.0.0"
    },
    "paths": {
...
        "/v1/business-centers/search": {
            "get": {
                "tags": [
                    "BusinessCenters"
                ],
                "summary": "Search for business centers",
                "operationId": "2bb1e12f71da1c8a67fb9e7a69a18ad8",
                "parameters": [
                    {
                        "name": "min_cost",
                        "in": "query",
                        "description": "Minimum cost filter",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/AssignmentsEnum"
                        }
                    },
                    {
                        "name": "max_cost",
                        "in": "query",
                        "description": "Maximum cost filter",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 100000
                        }
                    },
                    {
                        "name": "min_area",
                        "in": "query",
                        "description": "Minimum area filter",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 0
                        }
                    },
                    {
                        "name": "max_area",
                        "in": "query",
                        "description": "Maximum area filter",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10000000
                        }
                    },
                    {
                        "name": "filter_type",
                        "in": "query",
                        "description": "Filter type: \"rent\" or \"sale\"",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "rent"
                        }
                    },
                    {
                        "name": "has_parking",
                        "in": "query",
                        "description": "Filter by availability of parking",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    },
                    {
                        "name": "is_furnished",
                        "in": "query",
                        "description": "Filter by furnished status",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": false
                        }
                    },
                    {
                        "name": "block_types[]",
                        "in": "query",
                        "description": "Filter by block types",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            },
                            "example": [
                                "office"
                            ]
                        }
                    },
                    {
                        "name": "assignments[]",
                        "in": "query",
                        "description": "Filter by assignments",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            },
                            "example": [
                                "lease"
                            ]
                        }
                    },
                    {
                        "name": "time_foot[]",
                        "in": "query",
                        "description": "Filter by metro time foot (distance to metro)",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            },
                            "example": [
                                5
                            ]
                        }
                    },
                    {
                        "name": "district_types[]",
                        "in": "query",
                        "description": "Filter by district types",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            },
                            "example": [
                                "downtown"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term for filtering by name or address",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "office building"
                        }
                    },
                    {
                        "name": "class_codes[]",
                        "in": "query",
                        "description": "Filter by class codes",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            },
                            "example": [
                                "A"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully fetched the business centers based on the search filters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/BusinessCenterResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid filter parameters"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred, please try again later"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/business-centers/coordinates": {
            "get": {
                "tags": [
                    "BusinessCenters"
                ],
                "summary": "Get BC coordinates",
                "operationId": "94612e838ca4004919c1c5eb2387544b",
                "responses": {
                    "200": {
                        "description": "Successfully fetched the business centers based on the search filters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "properties": {
                                                    "latitude": {
                                                        "type": "number",
                                                        "format": "float"
                                                    },
                                                    "longitude": {
                                                        "type": "number",
                                                        "format": "float"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid filter parameters"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred, please try again later"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/class-codes/{id}": {
            "get": {
                "tags": [
                    "ClassCode"
                ],
                "description": "Display the specified Item.",
                "operationId": "ee45179ef8a7867dd6bc00e80f7c13e5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AssignmentsEnum": {
                "type": "string",
                "enum": [
                    "\u0416\u0438\u043b\u043e\u0439 \u0434\u043e\u043c",
                    "\u041e\u0442\u0434\u0435\u043b\u044c\u043d\u043e\u0441\u0442\u043e\u044f\u0449\u0435\u0435 \u0437\u0434\u0430\u043d\u0438\u0435",
                    "\u0411\u0438\u0437\u043d\u0435\u0441-\u0446\u0435\u043d\u0442\u0440",
                    "\u0411\u0438\u0437\u043d\u0435\u0441-\u043f\u0430\u0440\u043a",
                    "\u0422\u043e\u0440\u0433\u043e\u0432\u044b\u0439 \u0446\u0435\u043d\u0442\u0440",
                    "\u0410\u0434\u043c\u0438\u043d\u0438\u0441\u0442\u0440\u0430\u0442\u0438\u0432\u043d\u043e\u0435 \u0437\u0434\u0430\u043d\u0438\u0435",
                    "\u041e\u0441\u043e\u0431\u043d\u044f\u043a",
                    "\u0416\u0438\u043b\u043e\u0439 \u043a\u043e\u043c\u043f\u043b\u0435\u043a\u0441",
                    "\u0422\u043e\u0440\u0433\u043e\u0432\u043e-\u043e\u0444\u0438\u0441\u043d\u044b\u0439 \u043a\u043e\u043c\u043f\u043b\u0435\u043a\u0441",
                    "\u041c\u043d\u043e\u0433\u043e\u0444\u0443\u043d\u043a\u0446\u0438\u043e\u043d\u0430\u043b\u044c\u043d\u044b\u0439 \u043a\u043e\u043c\u043f\u043b\u0435\u043a\u0441",
                    "\u041e\u0444\u0438\u0441\u043d\u043e-\u0436\u0438\u043b\u043e\u0439 \u043a\u043e\u043c\u043f\u043b\u0435\u043a\u0441",
                    "\u041a\u043e\u043c\u043f\u043b\u0435\u043a\u0441 \u043e\u0444\u0438\u0441\u043d\u044b\u0445 \u0437\u0434\u0430\u043d\u0438\u0439",
                    "\u041e\u0444\u0438\u0441\u043d\u043e-\u0441\u043a\u043b\u0430\u0434\u0441\u043a\u043e\u0439 \u043a\u043e\u043c\u043f\u043b\u0435\u043a\u0441",
                    "\u041f\u0440\u0438\u0441\u0442\u0440\u043e\u0439\u043a\u0430 \u043a \u0436\u0438\u043b\u043e\u043c\u0443 \u0434\u043e\u043c\u0443",
                    "\u0410\u043d\u0433\u0430\u0440",
                    "\u0413\u0430\u0440\u0430\u0436",
                    "\u0413\u0430\u0440\u0430\u0436\u043d\u044b\u0439 \u043a\u043e\u043c\u043f\u043b\u0435\u043a\u0441",
                    "\u0413\u043e\u0441\u0442\u0438\u043d\u0438\u0446\u0430",
                    "\u0418\u043c\u0443\u0449\u0435\u0441\u0442\u0432\u0435\u043d\u043d\u044b\u0439 \u043a\u043e\u043c\u043f\u043b\u0435\u043a\u0441",
                    "\u041a\u043e\u043c\u043f\u043b\u0435\u043a\u0441 \u043e\u0441\u043e\u0431\u043d\u044f\u043a\u043e\u0432",
                    "\u041c\u043d\u043e\u0433\u043e\u043a\u0432\u0430\u0440\u0442\u0438\u0440\u043d\u044b\u0439 \u0434\u043e\u043c",
                    "\u041e\u0442\u0435\u043b\u044c",
                    "\u041e\u0444\u0438\u0441\u043d\u043e-\u0433\u043e\u0441\u0442\u0438\u043d\u0438\u0447\u043d\u044b\u0439 \u043a\u043e\u043c\u043f\u043b\u0435\u043a\u0441",
                    "\u041f\u0440\u043e\u0438\u0437\u0432\u043e\u0434\u0441\u0442\u0432\u0435\u043d\u043d\u044b\u0439 \u043a\u043e\u0440\u043f\u0443\u0441",
                    "\u041f\u0440\u043e\u0438\u0437\u0432\u043e\u0434\u0441\u0442\u0432\u0435\u043d\u043d\u044b\u0439 \u043a\u043e\u043c\u043f\u043b\u0435\u043a\u0441",
                    "\u041f\u0440\u043e\u0438\u0437\u0432\u043e\u0434\u0441\u0442\u0432\u0435\u043d\u043e-\u0441\u043a\u043b\u0430\u0434\u0441\u043a\u043e\u0439 \u043a\u043e\u043c\u043f\u043b\u0435\u043a\u0441",
                    "\u0421\u043a\u043b\u0430\u0434",
                    "\u0421\u043a\u043b\u0430\u0434\u0441\u043a\u043e\u0439 \u043a\u043e\u043c\u043f\u043b\u0435\u043a\u0441",
                    "\u0421\u043f\u043e\u0440\u0442\u0438\u0432\u043d\u044b\u0439 \u043a\u043e\u043c\u043f\u043b\u0435\u043a\u0441",
                    "\u0424\u0438\u0437\u043a\u0443\u043b\u044c\u0442\u0443\u0440\u043d\u043e-\u043e\u0437\u0434\u043e\u0440\u043e\u0432\u0438\u0442\u0435\u043b\u044c\u043d\u044b\u0439 \u043a\u043e\u043c\u043f\u043b\u0435\u043a\u0441"
                ]
            },
...
            "BusinessCenterResource": {
                "properties": {
                    "id": {
                        "description": "Transform the resource into an array.",
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "eng_name": {
                        "type": "string"
                    },
                    "gross_boma_area": {
                        "type": "integer"
                    },
                    "gross_leasable_area": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/",
            "description": "Server parameters"
        }
    ]
}